feat(client): define better approach to handle with cache for backend
Use the revalidatePath() function to refresh the cached request for
the API based on the game day.

For example: If today is day 02 and the game cached is for the day 01
the cache will be revalidated. Doing this, the performance of the
application in frontend will be strongly enhanced and only one request
will be made for the API per day.