feat(client): avoid submit invalid answer
Remove the onSubmit handler because if player hit enter button, it
will submit form, even if there's nothing on the input or have an
invalid option.

Instead, move the "submit" of the answer for the onClick event on
button and keep it disabled until a valid answer is selected from
the options.

Other important change is on the search of the input, if the player
write anything that doesn't match with any options, all options will
be loaded, keep easy to player to select the team even if him typed
an error.