feat(client): create page to get previous game and new component card day game
The page to get previous game will be similar with main page with
today's game but there'a key difference: the server action.

While the main page will use the server action to get the today's
game, the previous game page will get a specific game and so, it needs
to create new components to get the response for the other action,
first component to do this was the card day game.