refactor(game): make game component generic and create hook for localstorage
Make the Game component generic, now it will receive the day as a
prop and based on it, will render the today's game or past games.

Also did a refactoring in LocalStorage methods, created a hook to
save and retrieve data, decoupling modules and isolating
responsabilities.

Create a component TodayGame to load in main page using the previous
Game component passing today date as prop for it.