feat(client): format tips based on plural or singular statements
When there's more than one statement to compose a tip, for example,
the top scorer of the team was two players with same quantity of
goals, the API will return a statement with an conector " e " between
them, for example: "Bruno e Gabriel". When this happens, the tips
must change its write to make sens for reader. The same happens with
top scorer agains to.